Blackbox: correct FROMDRAW() macro for C division semantics
Integer division in C rounds towards zero, so if you want it to
consistently round down you need to ensure that the arguments are
positive.  FROMDRAW() didn't do that, so clicks off the top and left
corners of the grid got treated as being in the top row or left column
(row and column 0) rather than ignored.

This commit fixes the macro so that it offsets its argument upward
before the division and compensates afterwards.