Keen: fix another misuse of dsf_canonify.
Chris Boyle points out that outline_block_structure has a comment
saying that we're supposed to have picked a square with a boundary to
its left. dsf_canonify no longer guarantees that, but dsf_minimal
does. Switch to using that throughout the function.

'keen --generate 10#12345 --print 5x2' failed an assertion before this
fix, and now doesn't.